refactor(domain): share persistence string normalization
