+++ crates/luban_domain/src/persistence/mod.rs
@@ -1,6 +1,7 @@
 mod fonts;
 mod load;
 mod save;
+mod strings;
 
 pub(crate) use load::apply_persisted_app_state;
 pub(crate) use save::to_persisted_app_state;
